requirements.txt
pymongo~=4.15.5
python-dotenv~=1.0.1
faker